+++ docs4/src/lib.rs
@@ -1,6 +1,11 @@
 // Dummy crate so rustdoc-links can resolve items from dependencies.
+pub use vantage_core;
+pub use vantage_dataset;
+pub use vantage_diorama;
 pub use vantage_expressions;
 pub use vantage_mongodb;
 pub use vantage_sql;
 pub use vantage_table;
 pub use vantage_types;
+pub use vantage_vista;
+pub use vantage_vista_factory;